feat(core): remove quick add button from PLP
