+++ .changeset/young-gorillas-hang.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Remove "Quick add" button in PLP for products that have options. Will now just show a button that links to the product.
