+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/checkbox-field/fragment.ts
