+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/date-field/index.tsx
