+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/multi-line-text-field/fragment.ts
