+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/multiple-choice-field/fragment.ts
