+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/number-field/index.tsx
