+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/quantity-field.tsx
