+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/fields/text-field/index.tsx
