+++ core/app/[locale]/(default)/product/[slug]/_components/product-form/use-product-form.ts
